cleaned up messes
