--- backend/requirements.txt
@@ -1,7 +1,0 @@
-Flask==2.2.3
-opencv-python==4.5.5.64
-numpy==1.23.0
-tensorflow==2.5.0
-flask
-opencv-python
-gunicorn
