+++ vercel.json
@@ -1,0 +1,10 @@
+{
+    "version": 2,
+    "builds": [
+      { "src": "api/motion.py", "use": "@vercel/python" }
+    ],
+    "routes": [
+      { "src": "/api/motion", "dest": "api/motion.py" }
+    ]
+  }
+
